Renaming Inbox components directory to the more appropriate Messages directory
